perf(dom): reserve attribute capacity
Growing an attribute array one slot at a time retains each copied prefix
in the arena. Reserve capacity within the existing node size; copy shared
or exact-sized arrays before reserving slots.

Matched release runs cut time by 14.7-38.3% for 100-1,000 insertions and
10.8-38.0% for XML parsing at those counts. One insertion costs 3% more.
HTML parsing controls stay within 1.2% of baseline. Ten retained trees
with 1,000 attributes add 0.94 MiB RSS instead of 118.4 MiB.

Publish the scaling and replacement controls through the shared suite
and CodSpeed. The docs include lxml's remaining large-attribute advantage.

+++ src/turbohtml/_c/dom/tree_internal.h
@@ -270,7 +270,7 @@ static inline th_node *node_new(th_tree *tree, enum th_node_type type) {
     if (located) {
         *node_loc(node) = NULL; /* filled by insert_element when the element has a source tag */
     }
-    node->type = type;
+    node->type = (uint8_t)type;
     node->atom = TH_TAG_UNKNOWN;
     return node;
 }
